Keep V100 FP16 LoRA merges and saves in FP32
Introduces a shared FP32-safe residual merge path for LoRA-family modules so FP16-base training with `lora_fp32_compute=true` keeps additive outputs finite and returns FP32 activations instead of casting back to FP16. Adds V100/sm_70-specific storage enforcement (`force_fp32_storage`) so adapter params/buffers and checkpoint tensors stay FP32 on the protected FP16 path, updates attention backend boundary casting for low-precision kernels, and expands bench/docs/tests to cover overflow behavior, dtype contracts, and save-dtype exceptions.

+++ tests/test_loha_module.py
@@ -237,12 +237,13 @@ def test_fp32_compute_casts_all_official_operands(monkeypatch):
         (module.hada_w1_a.float() @ module.hada_w1_b.float())
         * (module.hada_w2_a.float() @ module.hada_w2_b.float())
     ) * module.scale
-    expected = module.org_forward(x) + torch.nn.functional.linear(
-        x.float(), expected_delta
-    ).to(torch.float16)
+    expected = (
+        module.org_forward(x).float()
+        + torch.nn.functional.linear(x.float(), expected_delta).float()
+    )
     output = base(x)
 
-    assert output.dtype == torch.float16
+    assert output.dtype == torch.float32
     assert torch.allclose(output, expected, atol=1e-3, rtol=1e-3)
     assert seen_dtypes
     assert set(seen_dtypes) == {torch.float32}
